refactor(ui): split useEntryEditForm into focused sub-hooks
useEntryEditForm was 487 lines bundling four concerns: form state +
submit/cancel, async secret loading (password + protected custom
fields), custom-icon/favicon side effects, and query-cache invalidation.
The return type had 22 fields across three clusters, with EntryEditForm
destructuring 20 of them — the same shape that signaled the bundling.

Extracts two focused sub-hooks:

- useEntryFormSecrets({ entry, dbId, form }) — owns isLoadingSecrets,
  secretLoadError, retrySecretLoad and the cancel-aware load effect.
  The "store previous prop in state" reset (on dbId/entryId change)
  moves with the state it resets.
- useEntryFormIconActions({ entryId, dbId, form }) — owns
  isFetchingFavicon, isClearingCustomIcon, hasCustomIcon, canFetchFavicon,
  fetchFaviconFromUrl, clearCustomIcon, and the submit-time helpers
  applyCustomIconChange and maybeAutoFetchFavicon, plus the private
  refreshFaviconQueries cache invalidation.

useEntryEditForm composes both sub-hooks internally and keeps form
state, submit, cancel, password generation, and the dirty-tracking
notify effect. Submit calls applyCustomIconChange/maybeAutoFetchFavicon
on the icon sub-hook at the points where the icon side effects belong.

Both sub-hooks take `form` as input — they need form.setValue
(password / customIconUuid / protected custom fields) and the dirty-
field state. Coupling them to react-hook-form is unavoidable.

Public surface is identical — the same 22 fields in the same flat
shape. EntryEditForm.tsx is unchanged. No test changes needed (the
test suite covers EntryEditForm at the component level, not the hook
directly).
