+++ src/hooks/use-entry-edit-form.ts
@@ -1,17 +1,16 @@
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useRef } from "react";
 import { standardSchemaResolver } from "@hookform/resolvers/standard-schema";
-import { useQueryClient } from "@tanstack/react-query";
 import { ask } from "@tauri-apps/plugin-dialog";
 import { type Resolver, useForm, useWatch } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 import { toast } from "sonner";
-import { useAppPreferences } from "@/hooks/use-app-preferences";
+import { useEntryFormIconActions } from "@/hooks/use-entry-form-icon-actions";
+import { useEntryFormSecrets } from "@/hooks/use-entry-form-secrets";
 import { useEntryMutations } from "@/hooks/use-entry-mutations";
 import { useTags } from "@/hooks/use-tags";
 import { PASSWORD_GENERATOR_DEFAULTS } from "@/lib/constants";
 import { entryFormSchema, type EntryFormValues } from "@/lib/formTypes";
-import { queryKeys } from "@/lib/query-keys";
-import { database, entries as entriesApi, generator } from "@/lib/tauri";
+import { generator } from "@/lib/tauri";
 import type { Entry } from "@/lib/types";
 
 interface UseEntryEditFormOptions {
@@ -84,20 +83,13 @@ export function useEntryEditForm({
   onDirtyChange,
 }: UseEntryEditFormOptions) {
   const { t } = useTranslation();
-  const queryClient = useQueryClient();
   const isEditMode = Boolean(entry);
   const entryId = entry?.id ?? null;
   const entryRef = useRef<Entry | null | undefined>(entry);
   const groupIdRef = useRef(groupId);
-  const [isLoadingSecrets, setIsLoadingSecrets] = useState(isEditMode);
-  const [secretLoadError, setSecretLoadError] = useState<string | null>(null);
-  const [secretReloadToken, setSecretReloadToken] = useState(0);
-  const [isFetchingFavicon, setIsFetchingFavicon] = useState(false);
-  const [isClearingCustomIcon, setIsClearingCustomIcon] = useState(false);
 
   const { createEntry, updateEntry, moveEntry } = useEntryMutations(dbId);
   const { data: availableTags } = useTags(dbId);
-  const { preferences } = useAppPreferences();
 
   const form = useForm<EntryFormValues>({
     resolver: standardSchemaResolver(
@@ -105,6 +97,20 @@ export function useEntryEditForm({
     ) as Resolver<EntryFormValues>,
     defaultValues: getEntryFormDefaults(entry, groupId),
   });
+
+  const { isLoadingSecrets, secretLoadError, retrySecretLoad } =
+    useEntryFormSecrets({ entry, dbId, form });
+
+  const {
+    isFetchingFavicon,
+    isClearingCustomIcon,
+    hasCustomIcon,
+    canFetchFavicon,
+    fetchFaviconFromUrl,
+    clearCustomIcon,
+    applyCustomIconChange,
+    maybeAutoFetchFavicon,
+  } = useEntryFormIconActions({ entryId, dbId, form });
 
   useEffect(() => {
     entryRef.current = entry;
@@ -114,17 +120,6 @@ export function useEntryEditForm({
     groupIdRef.current = groupId;
   }, [groupId]);
 
-  // Reset state when the active entry changes. React-internal state (error,
-  // loading flag) is reset during render via the "store previous prop in
-  // state" pattern; the RHF form.reset stays in an effect since it's an
-  // external system update.
-  const [prevFormKey, setPrevFormKey] = useState({ dbId, entryId });
-  if (prevFormKey.dbId !== dbId || prevFormKey.entryId !== entryId) {
-    setPrevFormKey({ dbId, entryId });
-    setSecretLoadError(null);
-    setIsLoadingSecrets(Boolean(entryId));
-  }
-
   useEffect(() => {
     const currentEntry = entryRef.current ?? null;
     form.reset(getEntryFormDefaults(currentEntry, groupIdRef.current));
@@ -133,60 +128,6 @@ export function useEntryEditForm({
   useEffect(() => {
     onDirtyChange?.(form.formState.isDirty);
   }, [form.formState.isDirty, onDirtyChange]);
-
-  useEffect(() => {
-    const currentEntry = entry;
-    if (!currentEntry) return;
-    let cancelled = false;
-
-    async function loadSecrets(entryForLoad: Entry) {
-      try {
-        const [password, ...protectedValues] = await Promise.all([
-          entriesApi.getPassword(dbId, entryForLoad.id),
-          ...entryForLoad.customFieldMeta
-            .filter((m) => m.isProtected)
-            .map((m) =>
-              entriesApi
-                .getProtectedCustomField(dbId, entryForLoad.id, m.key)
-                .then((r) => ({ key: m.key, value: r.value }))
-            ),
-        ]);
-
-        if (cancelled) return;
-
-        form.setValue("password", password, { shouldDirty: false });
-        setSecretLoadError(null);
-
-        for (const pv of protectedValues) {
-          const fieldIndex = entryForLoad.customFieldMeta.findIndex(
-            (m) => m.key === pv.key
-          );
-          if (fieldIndex !== -1) {
-            form.setValue(`customFields.${fieldIndex}.value`, pv.value, {
-              shouldDirty: false,
-            });
-          }
-        }
-      } catch (error) {
-        if (!cancelled) {
-          const message =
-            error instanceof Error ? error.message : String(error);
-          setSecretLoadError(message);
-          toast.error(t("entries.toast.secretLoadFailed"));
-        }
-      } finally {
-        if (!cancelled) {
-          setIsLoadingSecrets(false);
-        }
-      }
-    }
-
-    void loadSecrets(currentEntry);
-
-    return () => {
-      cancelled = true;
-    };
-  }, [entry, dbId, form, secretReloadToken, t]);
 
   // Auto-generate password when creating a new entry
   useEffect(() => {
@@ -339,112 +280,8 @@ export function useEntryEditForm({
     }
   }
 
-  function retrySecretLoad() {
-    setSecretLoadError(null);
-    setIsLoadingSecrets(true);
-    setSecretReloadToken((prev) => prev + 1);
-  }
-
   function setGeneratedPassword(password: string) {
     form.setValue("password", password, { shouldDirty: true });
-  }
-
-  async function applyCustomIconChange(
-    targetEntryId: string,
-    nextUuid: string | null
-  ) {
-    try {
-      const changed = nextUuid
-        ? await entriesApi.setCustomIcon(dbId, targetEntryId, nextUuid)
-        : await entriesApi.clearCustomIcon(dbId, targetEntryId);
-      if (changed) {
-        await database.save(dbId);
-        await refreshFaviconQueries(targetEntryId);
-      }
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      toast.error(
-        t("entries.toast.customIconAssignFailed", { error: message })
-      );
-    }
-  }
-
-  async function fetchFaviconFromUrl() {
-    if (!entryId) return;
-    setIsFetchingFavicon(true);
-    try {
-      const outcome = await entriesApi.fetchFavicon(dbId, entryId, true);
-      if (outcome === "updated") {
-        await database.save(dbId);
-        const refreshed = await entriesApi.get(dbId, entryId);
-        form.setValue("customIconUuid", refreshed.customIconUuid ?? null, {
-          shouldDirty: false,
-        });
-        toast.success(t("entries.toast.faviconUpdated"));
-        await refreshFaviconQueries(entryId);
-      } else if (outcome === "unchanged") {
-        toast.success(t("entries.toast.faviconAlreadyUpToDate"));
-      } else {
-        toast.error(t("entries.toast.faviconNotFound"));
-      }
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      toast.error(t("entries.toast.faviconUpdateFailed", { error: message }));
-    } finally {
-      setIsFetchingFavicon(false);
-    }
-  }
-
-  async function clearCustomIcon() {
-    if (!entryId) return;
-    setIsClearingCustomIcon(true);
-    try {
-      const changed = await entriesApi.clearCustomIcon(dbId, entryId);
-      if (changed) {
-        await database.save(dbId);
-        form.setValue("customIconUuid", null, { shouldDirty: false });
-        toast.success(t("entries.toast.customIconCleared"));
-        await refreshFaviconQueries(entryId);
-      }
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      toast.error(t("entries.toast.customIconClearFailed", { error: message }));
-    } finally {
-      setIsClearingCustomIcon(false);
-    }
-  }
-
-  async function maybeAutoFetchFavicon(
-    targetEntryId: string,
-    urlValue: string
-  ) {
-    if (!preferences?.security.autoDownloadFavicons) return;
-    if (!urlValue.trim()) return;
-    try {
-      const outcome = await entriesApi.fetchFavicon(dbId, targetEntryId, false);
-      if (outcome === "updated") {
-        await database.save(dbId);
-        await refreshFaviconQueries(targetEntryId);
-      }
-    } catch {
-      // Keep save flow non-blocking for favicon fetch failures.
-    }
-  }
-
-  async function refreshFaviconQueries(targetEntryId: string) {
-    await Promise.all([
-      queryClient.invalidateQueries({
-        queryKey: queryKeys.database.customIcons(dbId),
-      }),
-      queryClient.invalidateQueries({
-        queryKey: queryKeys.entries.detail(dbId, targetEntryId),
-      }),
-      queryClient.invalidateQueries({
-        predicate: (query) =>
-          query.queryKey[0] === queryKeys.entries.all[0] &&
-          query.queryKey[1] === dbId,
-      }),
-    ]);
   }
 
   const isPending =
@@ -456,9 +293,6 @@ export function useEntryEditForm({
   const watchedUsername =
     useWatch({ control: form.control, name: "username" }) ?? "";
   const watchedUrl = useWatch({ control: form.control, name: "url" }) ?? "";
-  const watchedCustomIconUuid =
-    useWatch({ control: form.control, name: "customIconUuid" }) ?? null;
-  const isUrlDirty = Boolean(form.formState.dirtyFields.url);
 
   return {
     form,
@@ -479,8 +313,8 @@ export function useEntryEditForm({
     setGeneratedPassword,
     isFetchingFavicon,
     isClearingCustomIcon,
-    hasCustomIcon: Boolean(watchedCustomIconUuid),
-    canFetchFavicon: Boolean(entryId && watchedUrl.trim() && !isUrlDirty),
+    hasCustomIcon,
+    canFetchFavicon,
     fetchFaviconFromUrl,
     clearCustomIcon,
   };
